MULE-9461: Implement watcher message source in new File connector - only stasrt when in primary node
