+++ extensions/file/src/test/resources/directory-listener-config.xml
@@ -20,6 +20,6 @@
     </flow>
 
     <sub-flow name="onEvent">
-        <expression-component>org.mule.extension.file.DirectoryListenerTestCase.onMessage(message)</expression-component>
+        <expression-component>org.mule.extension.file.DirectoryListenerFunctionalTestCase.onMessage(message)</expression-component>
     </sub-flow>
 </mule>
